assets/inbuilt/transformers/java-springboot/requirements.txt
ipdb
pyyaml
xmltodict
grpcio
grpcio-tools